security: tlk_driver: Make a fastcall to check if ote call needs to be restarted
If TE_SMC_FC_HAS_NS_WORK is available, use it to check if the ote stdcall
should be retried unstead of counting stdcalls. If a response or
ns-callback request arrived after the stdcall checked for it, the ote
stdcall would not always be retried before the 10 second timeout expired.

Change-Id: I8884ba0d1696eec27d33adb65ab8182d30f1580b
Signed-off-by: Arve Hjønnevåg <arve@android.com>